@@ -1,11 +1,10 @@
 <template>
   <a-button-icon
     class="f-footer"
-    @click="$emit('click')"
   >
     <a-icon
       :name="icon"
-      class="a-icon f-footer"
+      class="f-footer"
       :class="iconClass"
     />
     <div class="a-label f-button-icon-footer">
