@@ -1,47 +1,35 @@
 <template>
-  <validation-observer>
-    <validation-provider
-      :name="label.toLowerCase()"
-      :rules="rules"
-      v-slot="{ errors }"
-    >
-      <m-input
-        type="datetime-local"
-        v-model="vModel"
-        :disabled="disabled"
-        :placeholder="label"
-        :error="errors.length > 0"
-        :assist="errors[0] || assist"
-      />
-    </validation-provider>
-  </validation-observer>
+  <m-input
+    type="datetime-local"
+    v-model="vModel"
+    :disabled="disabled"
+    :placeholder="label"
+    :error="isError"
+    :assist="errorMessage || assist"
+  />
 </template>
 
 <!-- USAGE EXAMPLE
   <m-field-datetime
     :label="$t('form.field.date')"
     v-model="date"
-    :rules="rules.date"
+    :rules="validationRules.date"
     :disabled="blockForm"
   />
 -->
 
 <script>
 import MInput from 'molecules/input';
-import { mixins } from 'mixins/base';
 import { DATE_FORMATS, getDate, displayDate } from 'utils/date';
+import { fieldValidationMixin, useFieldValidation } from 'plugins/validation/field';
+import { computed, toRefs } from 'vue';
 
 export default {
   name: 'm-field-datetime',
-  mixins: [mixins.vModel],
   components: { MInput },
   props: {
     disabled: Boolean,
     label: {
-      type: String,
-      default: '',
-    },
-    rules: {
       type: String,
       default: '',
     },
@@ -50,18 +38,26 @@ export default {
       default: '',
     },
   },
-  computed: {
-    vModel: {
+  mixins: [fieldValidationMixin],
+  setup: (props, context) => {
+    const { modelValue } = toRefs(props);
+
+    const vModel = computed({
       get () {
-        if (this.value) return displayDate.inFormat(this.value, DATE_FORMATS.YYYYMMDDTHHmm);
+        if (modelValue.value) return displayDate.inFormat(modelValue.value, DATE_FORMATS.YYYYMMDDTHHmm);
         else return null;
       },
       set (value) {
         if (value) {
-          this.$emit('input', getDate.fromFormat(value, DATE_FORMATS.YYYYMMDDTHHmm));
+          context.emit('update:modelValue', getDate.fromFormat(value, DATE_FORMATS.YYYYMMDDTHHmm));
         } else return null;
       },
-    },
+    });
+
+    return {
+      ...useFieldValidation(props, context, { vModel }),
+      vModel,
+    };
   },
 };
 </script>
