@@ -5,8 +5,7 @@
       @actionTriggered="onSavePosition"
       @cancel="$emit('cancel')"
     >
-      <template slot="text">{{ $t('page.admin.setMapPosition.content') }}</template>
-      <template slot="button-name">{{ $t('form.button.choose') }}</template>
+      <template v-slot:text>{{ $t('page.admin.setMapPosition.content') }}</template>
     </m-banner-map>
   </o-map>
 </template>
@@ -15,6 +14,7 @@
 import OMap from 'organisms/map';
 import MBannerMap from 'molecules/banner-map';
 import { map } from 'map';
+import { DEFAULT_EVENT_CONFIG } from 'config/event-config';
 
 export default {
   name: 'o-admin-set-map-position',
@@ -30,9 +30,9 @@ export default {
   },
   mounted () {
     map.panTo({
-      latitude: this.event.mapLatitude,
-      longitude: this.event.mapLongitude,
-      zoom: this.event.mapZoom,
+      latitude: this.event.mapLatitude || DEFAULT_EVENT_CONFIG.newEvent.mapLatitude,
+      longitude: this.event.mapLongitude || DEFAULT_EVENT_CONFIG.newEvent.mapLongitude,
+      zoom: this.event.mapZoom || DEFAULT_EVENT_CONFIG.newEvent.mapZoom,
     });
   },
   methods: {
