@@ -1,5 +1,4 @@
-import Vue from 'vue';
-import Router from 'vue-router';
+import { createRouter, createWebHistory/*, isNavigationFailure, NavigationFailureType */ } from 'vue-router';
 import { store } from 'store';
 import { api } from 'api';
 import { routes } from './routes';
@@ -11,32 +10,11 @@ import { guardsUtils } from 'src/router/guards';
 
 let firstRun = true;
 
-Vue.use(Router);
-
-const router = new Router({
-  mode: 'history',
+const router = createRouter({
   base: process.env.BASE_URL,
   routes,
+  history: createWebHistory(),
 });
-
-(function silenceNavigationFailureErrors () {
-  // Solution from issue below, it removes ugly navigation duplicated error
-  // https://github.com/vuejs/vue-router/issues/2881#issuecomment-520554378
-  const originalPush = Router.prototype.push;
-  Router.prototype.push = function push (location, onResolve, onReject) {
-    if (onResolve || onReject) {
-      return originalPush.call(this, location, onResolve, onReject);
-    }
-    return originalPush.call(this, location).catch((err) => {
-      if (Router.isNavigationFailure(err, Router.NavigationFailureType.duplicated)) {
-        // resolve err
-        return err;
-      }
-      // rethrow error
-      return Promise.reject(err);
-    });
-  };
-})();
 
 router.beforeEach((to, from, next) => {
   let promise;
