@@ -1,5 +1,5 @@
 const webpack = require('webpack');
-const VueLoaderPlugin = require('vue-loader/lib/plugin');
+const { VueLoaderPlugin } = require('vue-loader');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const CopyPlugin = require('copy-webpack-plugin');
 const webpackUtils = require('./webpack/utils');
@@ -39,17 +39,19 @@ module.exports = {
   },
   resolve: {
     alias: {
-      vue: 'vue/dist/vue.esm.js',
       src: resolve('src'),
+      components: resolve('src/components'),
       api: resolve('src/api'),
       map: resolve('src/map'),
       store: resolve('src/store'),
       utils: resolve('src/utils'),
+      dictionary: resolve('src/dictionary'),
       vendors: resolve('../vendors'),
       validateCodes: resolve('../lib/validateCodes.js'),
       config: resolve('src/config'),
       models: resolve('src/models'),
 
+      plugins: resolve('src/components/plugins'),
       atoms: resolve('src/components/atoms'),
       extends: resolve('src/components/extends'),
       mixins: resolve('src/components/mixins'),
@@ -62,7 +64,7 @@ module.exports = {
   },
   plugins: [
     new ESLintPlugin({
-      extensions: ['js', 'vue'],
+      extensions: ['js', 'vue', 'ts'],
       formatter: require.resolve('eslint-friendly-formatter'),
       eslintPath: require.resolve('eslint'),
       useEslintrc: true,
