@@ -32,8 +32,11 @@ module.exports = {
     },
     {
       test: /\.ts?$/,
-      use: 'ts-loader',
+      loader: 'ts-loader',
       exclude: /node_modules/,
+      options: {
+        appendTsSuffixTo: [/\.vue$/],
+      },
     },
   ],
 };
